refactor(cas): drop buildConversationalAgentHeaders from public barrel
