+++ src/services/conversational-agent/helpers/index.ts
@@ -24,6 +24,3 @@ export * from './content-part-helper';
 
 // Transformers
 export * from './transformers';
-
-// Header builder
-export * from './header';
